+++ txpool/src/lib.rs
@@ -2,4 +2,7 @@ use miden_objects::transaction::ProvenTransaction;
 
 pub mod tasks;
 
+#[cfg(test)]
+pub mod test_utils;
+
 pub type TxBatch = Vec<ProvenTransaction>;
